OA-3 Alteração no tipo de valor de resposta de inteiro e grid de inteiros. De Integer para Long

+++ otus-activity/src/main/java/org/ccem/otus/model/survey/activity/filling/answer/IntegerAnswer.java
@@ -7,9 +7,9 @@ import org.ccem.otus.model.survey.activity.filling.AnswerFill;
 
 public class IntegerAnswer extends AnswerFill {
 
-	private Integer value;
+	private Long value;
 
-	public Integer getValue() {
+	public Long getValue() {
 		return value;
 	}
